chore: move build system to tsup
