+++ .vscode/launch.json
@@ -1,7 +1,3 @@
-// A launch configuration that compiles the extension and then opens it inside a new window
-// Use IntelliSense to learn about possible attributes.
-// Hover to view descriptions of existing attributes.
-// For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
 {
   "version": "0.2.0",
   "configurations": [
@@ -15,7 +11,7 @@
         "--folder-uri=${workspaceRoot}/packages/playground"
       ],
       "outFiles": [
-        "${workspaceFolder}/packages/vscode-ext/out/**/*.js"
+        "${workspaceFolder}/packages/vscode-ext/dist/**/*.js"
       ]
     },
     {
@@ -25,7 +21,7 @@
       "port": 6009,
       "restart": true,
       "outFiles": [
-        "${workspaceRoot}/*/*/out/**/*.js"
+        "${workspaceRoot}/*/*/dist/**/*.js"
       ]
     }
   ]
