+++ package.json
@@ -13,16 +13,13 @@
     "dev:compiler": "cross-env NODE_ENV=development pnpm --filter @vue-vine/compiler run dev",
     "dev:vite": "cross-env NODE_ENV=development pnpm --filter @vue-vine/vite-plugin run dev",
     "dev:main": "cross-env NODE_ENV=development pnpm --filter vue-vine run dev",
+    "dev:ext": "cross-env NODE_ENV=development esno scripts/ext-dev.js",
     "build:compiler": "cross-env NODE_ENV=production pnpm --filter @vue-vine/compiler run build",
     "build:vite": "cross-env NODE_ENV=production pnpm --filter @vue-vine/vite-plugin run build",
     "build:ls": "cross-env NODE_ENV=production pnpm --filter @vue-vine/language-server run build",
     "build:eslint-parser": "cross-env NODE_ENV=production pnpm --filter @vue-vine/eslint-parser run build",
     "build:main": "cross-env NODE_ENV=production pnpm --filter vue-vine run build",
-    "ext:dev": "cross-env NODE_ENV=development esno scripts/ext-dev.js",
-    "ext:tsc": "cross-env NODE_ENV=production pnpm --filter vue-vine-extension run build:tsc",
-    "ext:esbuild": "cross-env NODE_ENV=production pnpm --filter vue-vine-extension run build:esbuild",
-    "build:ext": "cross-env NODE_ENV=production pnpm ext:tsc && pnpm ext:esbuild",
-    "ext:build-full": "cross-env NODE_ENV=production pnpm build:base && pnpm ext:tsc && pnpm ext:esbuild",
+    "build:ext": "cross-env NODE_ENV=production pnpm --filter vue-vine-extension run build",
     "test": "esno scripts/run-test.js",
     "test:compiler": "pnpm --filter @vue-vine/compiler run test",
     "test:e2e": "pnpm --filter @vue-vine/e2e-test run test",
@@ -39,8 +36,11 @@
   },
   "devDependencies": {
     "@antfu/eslint-config": "^2.8.1",
+    "@antfu/ni": "^0.21.12",
     "@baiwusanyu/utils-log": "^1.0.15",
     "@changesets/cli": "^2.26.2",
+    "@types/node": "^20.11.27",
+    "@vitest/coverage-v8": "^0.33.0",
     "@vue-vine/eslint-parser": "workspace:*",
     "concurrently": "^8.2.1",
     "cross-env": "^7.0.3",
@@ -49,14 +49,15 @@
     "esno": "^0.16.3",
     "lint-staged": "^13.3.0",
     "playwright-chromium": "^1.37.1",
-    "rimraf": "^5.0.1",
-    "rollup": "^3.28.1",
     "shelljs": "^0.8.5",
     "simple-git-hooks": "^2.9.0",
     "tree-kill": "^1.2.2",
     "tslib": "^2.6.2",
+    "tsup": "^8.0.2",
     "typescript": "^5.2.2",
-    "vitest": "^1.3.1"
+    "vite": "^4.4.9",
+    "vitest": "^1.3.1",
+    "vue": "^3.3.4"
   },
   "pnpm": {
     "overrides": {
