+++ packages/compiler/package.json
@@ -10,7 +10,7 @@
   ],
   "exports": {
     ".": {
-      "dev": "./index.ts",
+      "dev": "./src/index.ts",
       "types": "./dist/index.d.ts",
       "import": "./dist/index.mjs",
       "require": "./dist/index.js"
@@ -23,8 +23,8 @@
     "dist"
   ],
   "scripts": {
-    "dev": "rollup -w -c ./rollup.config.mjs",
-    "build": "rollup -c ./rollup.config.mjs",
+    "dev": "tsup --watch",
+    "build": "tsup",
     "test": "vitest",
     "test:verbose": "VINE_DEV_VITEST=true vitest"
   },
@@ -42,18 +42,9 @@
     "source-map-js": "^1.0.2"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^25.0.4",
-    "@rollup/plugin-node-resolve": "^15.2.1",
     "@types/hash-sum": "^1.0.0",
     "@types/line-column": "^1.0.0",
-    "@vitest/coverage-v8": "^0.33.0",
-    "prettier": "^3.0.3",
-    "rollup": "^3.28.1",
-    "rollup-plugin-esbuild": "^5.0.0",
-    "typescript": "^5.2.2",
-    "vite": "^4.4.9",
-    "vitest": "^0.33.0",
-    "vue": "^3.3.4"
+    "prettier": "^3.0.3"
   },
   "publishConfig": {
     "access": "public",
