+++ packages/eslint-parser/package.json
@@ -10,8 +10,7 @@
   ],
   "exports": {
     ".": {
-      "dev": "./index.ts",
-      "types": "./dist/index.d.ts",
+      "dev": "./src/index.ts",
       "import": "./dist/index.mjs",
       "require": "./dist/index.js"
     },
@@ -19,14 +18,12 @@
   },
   "main": "./dist/index.js",
   "module": "./dist/index.mjs",
-  "types": "./dist/index.d.ts",
   "files": [
     "dist"
   ],
   "scripts": {
-    "dev": "rollup -w -c ./rollup.config.mjs",
-    "build": "rollup -c ./rollup.config.mjs",
-    "test": "vitest"
+    "dev": "tsup --watch",
+    "build": "tsup"
   },
   "dependencies": {
     "@types/lodash": "^4.14.201",
@@ -41,19 +38,11 @@
     "semver": "^7.5.4"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^25.0.7",
-    "@rollup/plugin-json": "^6.0.1",
-    "@rollup/plugin-node-resolve": "^15.2.3",
     "@types/debug": "^4.1.12",
+    "@types/eslint-scope": "^3.7.7",
     "@types/line-column": "^1.0.2",
     "@types/node": "^20.9.2",
     "@types/semver": "^7.5.5",
-    "@typescript-eslint/types": "^6.21.0",
-    "@vitest/coverage-v8": "^0.33.0",
-    "eslint": "^8.54.0",
-    "rollup": "^3.29.4",
-    "rollup-plugin-esbuild": "^5.0.0",
-    "typescript": "^5.2.2",
-    "vitest": "^0.34.6"
+    "eslint": "^8.54.0"
   }
 }
