+++ packages/language-server/tsup.config.ts
@@ -1,0 +1,1 @@
+export { default } from '../../tsup.config'
