+++ packages/vscode-ext/package.json
@@ -11,7 +11,7 @@
   "categories": [
     "Programming Languages"
   ],
-  "main": "./out/extension.js",
+  "main": "./dist/client.js",
   "icon": "assets/ext-icon.png",
   "engines": {
     "vscode": "^1.78.0"
@@ -67,10 +67,8 @@
   },
   "scripts": {
     "vscode:prepublish": "pnpm run compile",
-    "watch:tsc": "tsc -w --preserveWatchOutput",
-    "build:tsc": "tsc",
-    "watch:esbuild": "cross-env NODE_ENV=development node ./scripts/esbuild.js --watch",
-    "build:esbuild": "cross-env NODE_ENV=production node ./scripts/esbuild.js --minify",
+    "dev": "cross-env NODE_ENV=development tsup --watch",
+    "build": "cross-env NODE_ENV=production tsup --minify",
     "pretest": "pnpm run compile && pnpm run lint",
     "lint": "eslint src --ext ts"
   },
