+++ packages/vue-vine/tsconfig.json
@@ -1,23 +1,7 @@
 {
-  "compilerOptions": {
-    "target": "ESNext",
-    "customConditions": ["dev"],
-    "module": "ESNext",
-    "moduleResolution": "bundler",
-    "strict": true,
-    "declaration": true,
-    "declarationDir": "dist",
-    "emitDeclarationOnly": true,
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "skipLibCheck": true
-  },
+  "extends": "../../tsconfig.json",
   "include": [
     "src",
     "../../types/shims.d.ts"
-  ],
-  "exclude": [
-    "node_modules",
-    "dist"
   ]
 }
