+++ scripts/run-dev.js
@@ -1,5 +1,6 @@
 import { cliExec } from './utils'
 
+// TODO
 const devCommand = 'cross-env NODE_ENV=development pnpm concurrently '
   + '-p "  {name}  " '
   + '-n "COMPILER,VITE,MAIN" '
